finalizing SimpleGrid
