+++ src/components/Grid/SimpleGrid.stories.mdx
@@ -54,7 +54,7 @@ import DSProvider from "../../theme/provider";
 
 <Description of={SimpleGrid} />
 
-Donec id elit non mi porta gravida at eget metus. Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras mattis consectetur purus sit amet fermentum. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Vestibulum id ligula porta felis euismod semper.
+The `SimpleGrid` component is used to render UI elements in a uniform grid layout, with the column widths and grid spacing fixed throughout the entire grid.
 
 <Canvas withToolbar>
   <Story name="Basic SimpleGrid">
@@ -133,7 +133,7 @@ Donec id elit non mi porta gravida at eget metus. Lorem ipsum dolor sit amet, co
 
 <ArgsTable story="Basic SimpleGrid" />
 
-## Other Grid Examples
+## Other SimpleGrid Examples
 
 <Canvas>
   <Story name="Other SimpleGrid Examples">
